+++ java/src/main/java/com/techelevator/Vendable.java
@@ -4,7 +4,7 @@ public interface Vendable {
 
 	public String getSlotNumber();
 	public String getName();
-	public int getCost();
+	public double getCost();
 	public int getQuantity();
 	public void decrementQuantity();
 	public void printMessage();
